fix: Resolve No HTTP Trigger issue
- Moved `CosmosClient` initialization inside the serverless function to solve the No HTTP Trigger issues on production.
