+++ tsconfig.json
@@ -1,12 +1,10 @@
 {
   "compilerOptions": {
-   "target": "ES6", 
-   "module": "NodeNext",
+    "module": "commonjs",
+    "target": "es6",
     "outDir": "dist",
     "rootDir": ".",
     "sourceMap": true,
-    "strict": false,
-    "skipLibCheck": true,
-    "moduleResolution": "nodenext" // or "nodenext"
+    "strict": false
   }
 }
